JVMCBC-526 Compression ratio calculation should use floating point arithmetic
Motivation
==========
Compression ratio calculation currently divides two integers,
which will yield an integer result that is likely to be zero,
defeating the purpose of the check.

Modifications
=============
Cast one of the arguments to a double.

Result
======
Ratio calculation gives expected result.

Reviewed-on: http://review.couchbase.org/92730
Tested-by: Build Bot <build@couchbase.com>

+++ src/main/java/com/couchbase/client/core/endpoint/kv/KeyValueHandler.java
@@ -338,7 +338,7 @@ public class KeyValueHandler
             throw new RuntimeException("Could not snappy-compress value.", ex);
         }
 
-        double ratio = compressedContent.readableBytes() / uncompressedLength;
+        double ratio = (double) compressedContent.readableBytes() / uncompressedLength;
         if (ratio > minCompressionRatio) {
             // compressed is not smaller per ratio, so just send the original
             compressedContent.release();
